requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-crispy-forms==1.8.1
django-filter==2.2.0
idna==2.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.0
urllib3==1.25.8
